requirements.txt
Flask
gunicorn
Flask-Cors
pandas
numpy
PyYAML
zip-files
six
sklearn
-e .